refactor: consolidate platform-specific code and remove 275 unnecessary feature guards
This commit performs a major architectural cleanup to simplify cross-platform builds
and eliminate scattered conditional compilation guards.

## Platform-Specific Dependencies

Moved HDF5 and tracing dependencies to Unix-only (Linux/macOS):
- hdf5-metno and hdf5-metno-sys now under [target.'cfg(unix)'.dependencies]
- tracing and tracing-subscriber now under [target.'cfg(unix)'.dependencies]
- Windows builds automatically exclude these dependencies (no build errors)

## HDF5 Code Consolidation

Moved all HDF5 functionality to src/ev_formats/hdf5_reader.rs:
- load_events_from_hdf5() - 333 lines (main public API)
- call_python_prophesee_fallback() - 117 lines (Python interop)
- try_rust_ecf_decoder() - 31 lines (ECF decoder fallback)

File protected by single #![cfg(feature = "hdf5")] guard, eliminating need for
scattered Windows guards throughout codebase.

## Terminal Code Consolidation

Created src/ev_visualization/terminal_python.rs for terminal Python bindings:
- PyTerminalVisualizationConfig - configuration wrapper
- PyTerminalEventVisualizer - visualiser wrapper
- create_terminal_event_viewer() - factory function

File protected by single #![cfg(feature = "terminal")] guard, following same
pattern as HDF5 code consolidation.

## Feature Guard Removal

Removed ~275 unnecessary feature guards across 33 files:
- 195 #[cfg(feature = "polars")] guards removed
- 38 #[cfg(feature = "python")] guards removed
- 35 #[cfg(feature = "arrow")] guards removed
- 7 #[cfg(feature = "terminal")] guards removed (moved to dedicated file)

Rationale: polars, python, and arrow are default features that all evlib users require.

## Default Features Update

Updated default features in Cargo.toml:
- Added: arrow (zero-copy data transfer)
- Kept optional: terminal (not required by default)

Final defaults: ["polars", "python", "tracing", "hdf5", "arrow"]

## Dead Code Removal

- Removed candle-core related code (tensor_to_image function, ~99 lines)
- Removed duplicate Arrow stub implementations from arrow_builder.rs
- Removed conflicting From trait implementation (covered by stdlib blanket impl)

## Platform Compatibility

Linux/macOS: All features available (including HDF5 and tracing)
Windows: HDF5 and tracing automatically excluded, all other features work
All platforms: Terminal feature available as optional

## Impact

Net reduction: ~1,300 lines
Files modified: 33
Cleaner codebase with single guard points for platform-specific and optional code
