+++ src/ev_formats/aedat_reader.rs
@@ -1,5 +1,6 @@
 /// Comprehensive AEDAT format readers for all versions (1.0, 2.0, 3.1, 4.0)
 ///
+///
 /// This module provides readers for all AEDAT (Address Event Data) formats used by
 /// event cameras, with proper binary parsing, endianness handling, and error validation.
 ///
@@ -15,14 +16,11 @@
 /// - AEDAT File Format specifications
 use crate::ev_formats::dataframe_builder::EventDataFrameBuilder;
 use crate::ev_formats::EventFormat;
+use polars::prelude::*;
 use std::collections::HashMap;
 use std::fs::File;
 use std::io::{Read, Seek, SeekFrom};
 use std::path::Path;
-
-#[cfg(feature = "polars")]
-use polars::prelude::*;
-
 /// Errors that can occur during AEDAT reading
 #[derive(Debug)]
 pub enum AedatError {
@@ -58,7 +56,6 @@ pub enum AedatError {
         polarity: i8,
     },
 }
-
 impl std::fmt::Display for AedatError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
@@ -111,7 +108,6 @@ impl std::fmt::Display for AedatError {
         }
     }
 }
-
 impl std::error::Error for AedatError {
     fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
         match self {
@@ -120,13 +116,11 @@ impl std::error::Error for AedatError {
         }
     }
 }
-
 impl From<std::io::Error> for AedatError {
     fn from(error: std::io::Error) -> Self {
         AedatError::Io(error)
     }
 }
-
 /// AEDAT format version enum
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum AedatVersion {
@@ -135,7 +129,6 @@ pub enum AedatVersion {
     V3_1,
     V4_0,
 }
-
 impl std::fmt::Display for AedatVersion {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
@@ -146,7 +139,6 @@ impl std::fmt::Display for AedatVersion {
         }
     }
 }
-
 /// Metadata extracted from AEDAT headers
 #[derive(Debug, Clone, Default)]
 pub struct AedatMetadata {
@@ -169,7 +161,6 @@ pub struct AedatMetadata {
     /// Additional format-specific properties
     pub properties: HashMap<String, String>,
 }
-
 /// Configuration for AEDAT reading
 #[derive(Debug, Clone)]
 pub struct AedatConfig {
@@ -186,7 +177,6 @@ pub struct AedatConfig {
     /// Maximum number of events to read (None for all)
     pub max_events: Option<usize>,
 }
-
 impl Default for AedatConfig {
     fn default() -> Self {
         Self {
@@ -199,12 +189,10 @@ impl Default for AedatConfig {
         }
     }
 }
-
 /// Main AEDAT reader struct
 pub struct AedatReader {
     config: AedatConfig,
 }
-
 impl AedatReader {
     /// Create a new AEDAT reader with default configuration
     pub fn new() -> Self {
@@ -212,12 +200,10 @@ impl AedatReader {
             config: AedatConfig::default(),
         }
     }
-
     /// Create a new AEDAT reader with custom configuration
     pub fn with_config(config: AedatConfig) -> Self {
         Self { config }
     }
-
     /// Read AEDAT file and return events with metadata
     pub fn read_file<P: AsRef<Path>>(
         &self,
@@ -225,7 +211,6 @@ impl AedatReader {
     ) -> Result<(DataFrame, AedatMetadata), AedatError> {
         let mut file = File::open(path.as_ref())?;
         let version = self.detect_version(&mut file)?;
-
         match version {
             AedatVersion::V1_0 => self.read_aedat_1_0(&mut file),
             AedatVersion::V2_0 => self.read_aedat_2_0(&mut file),
@@ -233,63 +218,52 @@ impl AedatReader {
             AedatVersion::V4_0 => self.read_aedat_4_0(&mut file),
         }
     }
-
     /// Detect AEDAT version from file magic bytes
     fn detect_version(&self, file: &mut File) -> Result<AedatVersion, AedatError> {
         let mut buffer = [0u8; 32];
         file.seek(SeekFrom::Start(0))?;
         let bytes_read = file.read(&mut buffer)?;
-
         if bytes_read < 10 {
             return Err(AedatError::InsufficientData {
                 expected: 10,
                 actual: bytes_read,
             });
         }
-
         // Check for version magic bytes - order matters!
         if buffer.starts_with(b"AEDAT4") {
             return Ok(AedatVersion::V4_0);
         }
-
         if buffer.starts_with(b"#!AER-DAT2.0") {
             return Ok(AedatVersion::V2_0);
         }
-
         if buffer.starts_with(b"#!AER-DAT1.0") {
             return Ok(AedatVersion::V1_0);
         }
-
         // Check for 3.1 after checking for specific versions
         if buffer.starts_with(b"#!AER-DAT3") {
             return Ok(AedatVersion::V3_1);
         }
-
         if buffer.starts_with(b"#!AER-DAT") {
             // Generic AER-DAT header, default to 3.1
             return Ok(AedatVersion::V3_1);
         }
-
         // Try to infer version from file structure
         // AEDAT 1.0 often has no header or minimal header
         // Check if file starts with binary data pattern
         if self.looks_like_aedat_1_0(&buffer) {
             return Ok(AedatVersion::V1_0);
         }
-
         let first_bytes = &buffer[..bytes_read.min(32)];
         Err(AedatError::InvalidVersion(format!(
             "Unknown AEDAT version, first 32 bytes: {first_bytes:?}"
         )))
     }
-
     /// Check if data looks like AEDAT 1.0 format
     fn looks_like_aedat_1_0(&self, buffer: &[u8]) -> bool {
         // AEDAT 1.0 has 6-byte events
         if buffer.len() < 6 {
             return false;
         }
-
         // Check if data could be 6-byte aligned events
         // Very basic heuristic - check if we have reasonable coordinate values
         for chunk in buffer.chunks_exact(6) {
@@ -297,41 +271,32 @@ impl AedatReader {
                 // Try to parse as 16-bit address + 32-bit timestamp
                 let address = u16::from_le_bytes([chunk[0], chunk[1]]);
                 let _timestamp = u32::from_le_bytes([chunk[2], chunk[3], chunk[4], chunk[5]]);
-
                 // Extract coordinates from address (assuming 9-bit x, 9-bit y)
                 let x = (address >> 1) & 0x1FF;
                 let y = (address >> 10) & 0x1FF;
-
                 // Check if coordinates are reasonable
                 if x < 1024 && y < 1024 {
                     return true;
                 }
             }
         }
-
         false
     }
-
     /// Read AEDAT 1.0 format
     fn read_aedat_1_0(&self, file: &mut File) -> Result<(DataFrame, AedatMetadata), AedatError> {
         let mut metadata = AedatMetadata {
             version: Some(AedatVersion::V1_0),
             ..Default::default()
         };
-
         // Check for optional header
         file.seek(SeekFrom::Start(0))?;
         let header_size = self.parse_aedat_1_0_header(file, &mut metadata)?;
-
         // Read binary event data
         file.seek(SeekFrom::Start(header_size))?;
         let events = self.read_aedat_1_0_events(file)?;
-
         metadata.event_count = Some(events.height());
-
         Ok((events, metadata))
     }
-
     /// Parse AEDAT 1.0 header (optional)
     fn parse_aedat_1_0_header(
         &self,
@@ -341,31 +306,24 @@ impl AedatReader {
         file.seek(SeekFrom::Start(0))?;
         let mut buffer = [0u8; 1024];
         let bytes_read = file.read(&mut buffer)?;
-
         if bytes_read == 0 {
             return Ok(0);
         }
-
         // Check if file starts with a header
         if buffer.starts_with(b"#!AER-DAT1.0") {
             // Find the end of header by looking for the first non-header line
             let mut header_end = 0;
             let mut current_pos = 0;
-
             while current_pos < bytes_read {
                 let line_start = current_pos;
-
                 // Find end of line
                 while current_pos < bytes_read && buffer[current_pos] != b'\n' {
                     current_pos += 1;
                 }
-
                 if current_pos < bytes_read {
                     current_pos += 1; // Skip newline
                 }
-
                 let line_bytes = &buffer[line_start..current_pos.min(bytes_read)];
-
                 // Check if this line starts with # (header line)
                 if line_bytes.starts_with(b"#") {
                     // Parse header information if it's valid UTF-8
@@ -382,7 +340,6 @@ impl AedatReader {
                                     Some((metadata.sensor_resolution.unwrap_or((0, 0)).0, height));
                             }
                         }
-
                         metadata
                             .properties
                             .insert(format!("header_line_{header_end}"), line.to_string());
@@ -393,37 +350,30 @@ impl AedatReader {
                     break;
                 }
             }
-
             metadata.header_size = header_end as u64;
             return Ok(header_end as u64);
         }
-
         // No header found, data starts at beginning
         metadata.header_size = 0;
         Ok(0)
     }
-
     /// Read AEDAT 1.0 events (6 bytes per event)
     fn read_aedat_1_0_events(&self, file: &mut File) -> Result<DataFrame, AedatError> {
-        #[cfg(feature = "polars")]
         {
             let mut builder = EventDataFrameBuilder::new(EventFormat::AEDAT1, 100_000);
             let mut buffer = [0u8; 6];
             let mut event_index = 0;
             let mut prev_timestamp = 0.0;
-
             loop {
                 if let Some(max_events) = self.config.max_events {
                     if event_index >= max_events {
                         break;
                     }
                 }
-
                 let bytes_read = file.read(&mut buffer)?;
                 if bytes_read == 0 {
                     break; // EOF
                 }
-
                 if bytes_read != 6 {
                     if self.config.skip_invalid_events {
                         continue;
@@ -434,11 +384,9 @@ impl AedatReader {
                         });
                     }
                 }
-
                 // Parse 16-bit address + 32-bit timestamp
                 let address = u16::from_le_bytes([buffer[0], buffer[1]]);
                 let timestamp = u32::from_le_bytes([buffer[2], buffer[3], buffer[4], buffer[5]]);
-
                 // Extract coordinates and polarity from address
                 // AEDAT 1.0 format for DVS128 (standard specification):
                 // 16-bit address field breakdown:
@@ -449,21 +397,16 @@ impl AedatReader {
                 //   - '1' = Increase (ON event)
                 //   - '0' = Decrease (OFF event)
                 // Coordinate system: (0,0) in lower left corner
-
                 // Extract polarity from bit 0
                 let polarity = (address & 1) == 1;
-
                 // Extract x coordinate from bits 7-1 (7 bits)
                 let x = (address >> 1) & 0x7F; // 7 bits: 0x7F = 0111 1111
-
-                // Extract y coordinate from bits 14-8 (7 bits)
+                                               // Extract y coordinate from bits 14-8 (7 bits)
                 let y = (address >> 8) & 0x7F; // 7 bits: 0x7F = 0111 1111
-
-                // Note: For DVS128, coordinates are max 128x128, so 7 bits each is sufficient
-                // The coordinate system has (0,0) at lower left, but we convert to upper left
-                // by inverting y coordinate if sensor resolution is known
-
-                // Validate coordinates and timestamp before adding to builder
+                                               // Note: For DVS128, coordinates are max 128x128, so 7 bits each is sufficient
+                                               // The coordinate system has (0,0) at lower left, but we convert to upper left
+                                               // by inverting y coordinate if sensor resolution is known
+                                               // Validate coordinates and timestamp before adding to builder
                 if self.config.validate_timestamps {
                     if x > 127 || y > 127 {
                         if self.config.skip_invalid_events {
@@ -478,7 +421,6 @@ impl AedatReader {
                             });
                         }
                     }
-
                     let current_timestamp = timestamp as f64;
                     if current_timestamp < prev_timestamp {
                         if self.config.skip_invalid_events {
@@ -493,45 +435,30 @@ impl AedatReader {
                     }
                     prev_timestamp = current_timestamp;
                 }
-
                 builder.add_event(x, y, timestamp as f64, polarity);
                 event_index += 1;
             }
-
             builder.build().map_err(|e| AedatError::InvalidBinaryData {
                 offset: 0,
                 message: format!("Failed to build DataFrame: {}", e),
             })
         }
-        #[cfg(not(feature = "polars"))]
-        {
-            Err(AedatError::InvalidBinaryData {
-                offset: 0,
-                message: "Polars feature not enabled".to_string(),
-            })
-        }
     }
-
     /// Read AEDAT 2.0 format
     fn read_aedat_2_0(&self, file: &mut File) -> Result<(DataFrame, AedatMetadata), AedatError> {
         let mut metadata = AedatMetadata {
             version: Some(AedatVersion::V2_0),
             ..Default::default()
         };
-
         // Parse header
         file.seek(SeekFrom::Start(0))?;
         let header_size = self.parse_aedat_2_0_header(file, &mut metadata)?;
-
         // Read binary event data
         file.seek(SeekFrom::Start(header_size))?;
         let events = self.read_aedat_2_0_events(file)?;
-
         metadata.event_count = Some(events.height());
-
         Ok((events, metadata))
     }
-
     /// Parse AEDAT 2.0 header
     fn parse_aedat_2_0_header(
         &self,
@@ -543,28 +470,23 @@ impl AedatReader {
         let mut temp_buffer = [0u8; 1024];
         let mut header_size = 0u64;
         let mut line_index = 0;
-
         // Read file in chunks to find header end
         loop {
             let bytes_read = file.read(&mut temp_buffer)?;
             if bytes_read == 0 {
                 break;
             }
-
             buffer.extend_from_slice(&temp_buffer[..bytes_read]);
-
             // Look for header end - either double newline or end of file
             if let Some(pos) = buffer.windows(2).position(|w| w == b"\n\n") {
                 header_size = pos as u64 + 2; // +2 to skip the double newline
                 break;
             }
-
             if bytes_read < 1024 {
                 header_size = buffer.len() as u64;
                 break;
             }
         }
-
         // Parse header lines
         let header_str = String::from_utf8_lossy(&buffer[..header_size as usize]);
         for line in header_str.lines() {
@@ -581,7 +503,6 @@ impl AedatReader {
                             Some((metadata.sensor_resolution.unwrap_or((0, 0)).0, height));
                     }
                 }
-
                 metadata
                     .properties
                     .insert(format!("header_line_{line_index}"), line.to_string());
@@ -591,32 +512,26 @@ impl AedatReader {
                 break;
             }
         }
-
         metadata.header_size = header_size;
         Ok(header_size)
     }
-
     /// Read AEDAT 2.0 events (big-endian 32-bit timestamp + address pairs)
     fn read_aedat_2_0_events(&self, file: &mut File) -> Result<DataFrame, AedatError> {
-        #[cfg(feature = "polars")]
         {
             let mut builder = EventDataFrameBuilder::new(EventFormat::AEDAT1, 100_000);
             let mut buffer = [0u8; 8]; // 4 bytes timestamp + 4 bytes address
             let mut event_index = 0;
             let mut prev_timestamp = 0.0;
-
             loop {
                 if let Some(max_events) = self.config.max_events {
                     if event_index >= max_events {
                         break;
                     }
                 }
-
                 let bytes_read = file.read(&mut buffer)?;
                 if bytes_read == 0 {
                     break; // EOF
                 }
-
                 if bytes_read != 8 {
                     if self.config.skip_invalid_events {
                         continue;
@@ -627,11 +542,9 @@ impl AedatReader {
                         });
                     }
                 }
-
                 // Parse big-endian 32-bit timestamp and address
                 let timestamp = u32::from_be_bytes([buffer[0], buffer[1], buffer[2], buffer[3]]);
                 let address = u32::from_be_bytes([buffer[4], buffer[5], buffer[6], buffer[7]]);
-
                 // Extract coordinates and polarity from address
                 // AEDAT 2.0 format typically uses standard AER encoding:
                 // 32-bit address field breakdown:
@@ -641,17 +554,13 @@ impl AedatReader {
                 // - Bits 1-15: X address coordinate (up to 15 bits)
                 // - Bits 16-30: Y address coordinate (up to 15 bits)
                 // - Bit 31: Reserved/unused
-
                 // Extract polarity from bit 0
                 let polarity = (address & 1) == 1;
-
                 // Extract x coordinate from bits 1-15 (up to 15 bits)
                 let x = (address >> 1) & 0x7FFF; // 15 bits: 0x7FFF = 0111 1111 1111 1111
-
-                // Extract y coordinate from bits 16-30 (up to 15 bits)
+                                                 // Extract y coordinate from bits 16-30 (up to 15 bits)
                 let y = (address >> 16) & 0x7FFF; // 15 bits: 0x7FFF = 0111 1111 1111 1111
-
-                // Validate coordinates and timestamp before adding to builder
+                                                  // Validate coordinates and timestamp before adding to builder
                 if self.config.validate_timestamps {
                     let current_timestamp = timestamp as f64;
                     if current_timestamp < prev_timestamp {
@@ -667,45 +576,30 @@ impl AedatReader {
                     }
                     prev_timestamp = current_timestamp;
                 }
-
                 builder.add_event(x as u16, y as u16, timestamp as f64, polarity);
                 event_index += 1;
             }
-
             builder.build().map_err(|e| AedatError::InvalidBinaryData {
                 offset: 0,
                 message: format!("Failed to build DataFrame: {}", e),
             })
         }
-        #[cfg(not(feature = "polars"))]
-        {
-            Err(AedatError::InvalidBinaryData {
-                offset: 0,
-                message: "Polars feature not enabled".to_string(),
-            })
-        }
     }
-
     /// Read AEDAT 3.1 format
     fn read_aedat_3_1(&self, file: &mut File) -> Result<(DataFrame, AedatMetadata), AedatError> {
         let mut metadata = AedatMetadata {
             version: Some(AedatVersion::V3_1),
             ..Default::default()
         };
-
         // Parse header
         file.seek(SeekFrom::Start(0))?;
         let header_size = self.parse_aedat_3_1_header(file, &mut metadata)?;
-
         // Read binary event data
         file.seek(SeekFrom::Start(header_size))?;
         let events = self.read_aedat_3_1_events(file)?;
-
         metadata.event_count = Some(events.height());
-
         Ok((events, metadata))
     }
-
     /// Parse AEDAT 3.1 header
     fn parse_aedat_3_1_header(
         &self,
@@ -717,28 +611,23 @@ impl AedatReader {
         let mut temp_buffer = [0u8; 1024];
         let mut header_size = 0u64;
         let mut line_index = 0;
-
         // Read file in chunks to find header end
         loop {
             let bytes_read = file.read(&mut temp_buffer)?;
             if bytes_read == 0 {
                 break;
             }
-
             buffer.extend_from_slice(&temp_buffer[..bytes_read]);
-
             // Look for header end - either double newline or end of file
             if let Some(pos) = buffer.windows(2).position(|w| w == b"\n\n") {
                 header_size = pos as u64 + 2; // +2 to skip the double newline
                 break;
             }
-
             if bytes_read < 1024 {
                 header_size = buffer.len() as u64;
                 break;
             }
         }
-
         // Parse header lines
         let header_str = String::from_utf8_lossy(&buffer[..header_size as usize]);
         for line in header_str.lines() {
@@ -755,7 +644,6 @@ impl AedatReader {
                             Some((metadata.sensor_resolution.unwrap_or((0, 0)).0, height));
                     }
                 }
-
                 metadata
                     .properties
                     .insert(format!("header_line_{line_index}"), line.to_string());
@@ -765,32 +653,26 @@ impl AedatReader {
                 break;
             }
         }
-
         metadata.header_size = header_size;
         Ok(header_size)
     }
-
     /// Read AEDAT 3.1 events (signed little-endian format)
     fn read_aedat_3_1_events(&self, file: &mut File) -> Result<DataFrame, AedatError> {
-        #[cfg(feature = "polars")]
         {
             let mut builder = EventDataFrameBuilder::new(EventFormat::AEDAT1, 100_000);
             let mut buffer = [0u8; 8]; // 4 bytes timestamp + 4 bytes address (little-endian)
             let mut event_index = 0;
             let mut prev_timestamp = 0.0;
-
             loop {
                 if let Some(max_events) = self.config.max_events {
                     if event_index >= max_events {
                         break;
                     }
                 }
-
                 let bytes_read = file.read(&mut buffer)?;
                 if bytes_read == 0 {
                     break; // EOF
                 }
-
                 if bytes_read != 8 {
                     if self.config.skip_invalid_events {
                         continue;
@@ -801,11 +683,9 @@ impl AedatReader {
                         });
                     }
                 }
-
                 // Parse little-endian 32-bit timestamp and address
                 let timestamp = u32::from_le_bytes([buffer[0], buffer[1], buffer[2], buffer[3]]);
                 let address = u32::from_le_bytes([buffer[4], buffer[5], buffer[6], buffer[7]]);
-
                 // Extract coordinates and polarity from address
                 // AEDAT 3.1 format specification (little-endian):
                 // 32-bit data field breakdown:
@@ -816,7 +696,6 @@ impl AedatReader {
                 // - Bits 2-16: Y event address (up to 15 bits)
                 // - Bits 17-31: X event address (up to 15 bits)
                 // Coordinate system: (0,0) in upper left corner
-
                 // Check validity bit (bit 0)
                 let is_valid = (address & 1) == 1;
                 if !is_valid {
@@ -829,17 +708,13 @@ impl AedatReader {
                         });
                     }
                 }
-
                 // Extract polarity from bit 1
                 let polarity = ((address >> 1) & 1) == 1;
-
                 // Extract y coordinate from bits 2-16 (up to 15 bits)
                 let y = (address >> 2) & 0x7FFF; // 15 bits: 0x7FFF = 0111 1111 1111 1111
-
-                // Extract x coordinate from bits 17-31 (up to 15 bits)
+                                                 // Extract x coordinate from bits 17-31 (up to 15 bits)
                 let x = (address >> 17) & 0x7FFF; // 15 bits: 0x7FFF = 0111 1111 1111 1111
-
-                // Validate coordinates and timestamp before adding to builder
+                                                  // Validate coordinates and timestamp before adding to builder
                 if self.config.validate_timestamps {
                     let current_timestamp = timestamp as f64;
                     if current_timestamp < prev_timestamp {
@@ -855,45 +730,30 @@ impl AedatReader {
                     }
                     prev_timestamp = current_timestamp;
                 }
-
                 builder.add_event(x as u16, y as u16, timestamp as f64, polarity);
                 event_index += 1;
             }
-
             builder.build().map_err(|e| AedatError::InvalidBinaryData {
                 offset: 0,
                 message: format!("Failed to build DataFrame: {}", e),
             })
         }
-        #[cfg(not(feature = "polars"))]
-        {
-            Err(AedatError::InvalidBinaryData {
-                offset: 0,
-                message: "Polars feature not enabled".to_string(),
-            })
-        }
     }
-
     /// Read AEDAT 4.0 format
     fn read_aedat_4_0(&self, file: &mut File) -> Result<(DataFrame, AedatMetadata), AedatError> {
         let mut metadata = AedatMetadata {
             version: Some(AedatVersion::V4_0),
             ..Default::default()
         };
-
         // Parse header
         file.seek(SeekFrom::Start(0))?;
         let header_size = self.parse_aedat_4_0_header(file, &mut metadata)?;
-
         // Read binary event data with DV framework packet structure
         file.seek(SeekFrom::Start(header_size))?;
         let events = self.read_aedat_4_0_events(file)?;
-
         metadata.event_count = Some(events.height());
-
         Ok((events, metadata))
     }
-
     /// Parse AEDAT 4.0 header (DV framework)
     fn parse_aedat_4_0_header(
         &self,
@@ -905,16 +765,13 @@ impl AedatReader {
         let mut temp_buffer = [0u8; 1024];
         let mut header_size = 0u64;
         let mut line_index = 0;
-
         // Read file in chunks to find header end
         loop {
             let bytes_read = file.read(&mut temp_buffer)?;
             if bytes_read == 0 {
                 break;
             }
-
             buffer.extend_from_slice(&temp_buffer[..bytes_read]);
-
             // Look for header end (look for first non-comment line after AEDAT4)
             let header_str = String::from_utf8_lossy(&buffer);
             let mut found_header_end = false;
@@ -937,17 +794,14 @@ impl AedatReader {
                 found_header_end = true;
                 break;
             }
-
             if found_header_end {
                 break;
             }
-
             if bytes_read < 1024 {
                 header_size = buffer.len() as u64;
                 break;
             }
         }
-
         // Parse header lines
         let header_str = String::from_utf8_lossy(&buffer[..header_size as usize]);
         for line in header_str.lines() {
@@ -967,7 +821,6 @@ impl AedatReader {
                             Some((metadata.sensor_resolution.unwrap_or((0, 0)).0, height));
                     }
                 }
-
                 metadata
                     .properties
                     .insert(format!("header_line_{line_index}"), line.to_string());
@@ -977,33 +830,27 @@ impl AedatReader {
                 break;
             }
         }
-
         metadata.header_size = header_size;
         Ok(header_size)
     }
-
     /// Read AEDAT 4.0 events (DV framework with 28-byte packet headers)
     fn read_aedat_4_0_events(&self, file: &mut File) -> Result<DataFrame, AedatError> {
-        #[cfg(feature = "polars")]
         {
             let mut builder = EventDataFrameBuilder::new(EventFormat::AEDAT1, 100_000);
             let mut event_index = 0;
             let mut prev_timestamp = 0.0;
-
             loop {
                 if let Some(max_events) = self.config.max_events {
                     if event_index >= max_events {
                         break;
                     }
                 }
-
                 // Read packet header (28 bytes)
                 let mut packet_header = [0u8; 28];
                 let bytes_read = file.read(&mut packet_header)?;
                 if bytes_read == 0 {
                     break; // EOF
                 }
-
                 if bytes_read != 28 {
                     if self.config.skip_invalid_events {
                         continue;
@@ -1014,7 +861,6 @@ impl AedatReader {
                         });
                     }
                 }
-
                 // Parse packet header to determine packet type and size
                 let packet_type = u16::from_le_bytes([packet_header[0], packet_header[1]]);
                 let packet_size = u32::from_le_bytes([
@@ -1023,12 +869,10 @@ impl AedatReader {
                     packet_header[6],
                     packet_header[7],
                 ]);
-
                 // Only process polarity event packets (type 1)
                 if packet_type == 1 {
                     let packet_events =
                         self.read_aedat_4_0_packet_events(file, packet_size as usize)?;
-
                     for (x, y, timestamp, polarity) in packet_events {
                         // Validate coordinates and timestamp before adding to builder
                         if self.config.validate_timestamps {
@@ -1045,7 +889,6 @@ impl AedatReader {
                             }
                             prev_timestamp = timestamp;
                         }
-
                         builder.add_event(x, y, timestamp, polarity);
                         event_index += 1;
                     }
@@ -1054,21 +897,12 @@ impl AedatReader {
                     file.seek(SeekFrom::Current(packet_size as i64))?;
                 }
             }
-
             builder.build().map_err(|e| AedatError::InvalidBinaryData {
                 offset: 0,
                 message: format!("Failed to build DataFrame: {}", e),
             })
         }
-        #[cfg(not(feature = "polars"))]
-        {
-            Err(AedatError::InvalidBinaryData {
-                offset: 0,
-                message: "Polars feature not enabled".to_string(),
-            })
-        }
     }
-
     /// Read events from AEDAT 4.0 packet
     fn read_aedat_4_0_packet_events(
         &self,
@@ -1078,45 +912,36 @@ impl AedatReader {
         let mut events = Vec::new();
         let mut buffer = vec![0u8; packet_size];
         let bytes_read = file.read(&mut buffer)?;
-
         if bytes_read != packet_size {
             return Err(AedatError::InsufficientData {
                 expected: packet_size,
                 actual: bytes_read,
             });
         }
-
         // Parse polarity events (8 bytes each in DV format)
         for chunk in buffer.chunks_exact(8) {
             let timestamp = u32::from_le_bytes([chunk[0], chunk[1], chunk[2], chunk[3]]);
             let x = u16::from_le_bytes([chunk[4], chunk[5]]);
             let y = u16::from_le_bytes([chunk[6], chunk[7]]);
-
             // AEDAT 4.0 DV framework format:
             // - Timestamp: 32-bit microsecond timestamp
             // - X coordinate: 16-bit value (0-65535)
             // - Y coordinate: 16-bit value with MSB encoding polarity
             //   - Bit 15: Polarity (1 = ON, 0 = OFF)
             //   - Bits 0-14: Y coordinate (0-32767)
-
             // Extract polarity from MSB of y
             let polarity = (y & 0x8000) != 0;
-
             // Extract y coordinate by masking out polarity bit
             let y_clean = y & 0x7FFF; // Remove polarity bit (bit 15)
-
-            // Validate coordinates are within reasonable bounds
+                                      // Validate coordinates are within reasonable bounds
             if x >= 8192 || y_clean >= 8192 {
                 // Skip events with unreasonable coordinates (likely corrupted)
                 continue;
             }
-
             events.push((x, y_clean, timestamp as f64, polarity));
         }
-
         Ok(events)
     }
-
     /// Extract numeric value from a header line
     fn extract_number_from_line(&self, line: &str) -> Option<u16> {
         let parts: Vec<&str> = line.split_whitespace().collect();
@@ -1128,34 +953,28 @@ impl AedatReader {
         None
     }
 }
-
 impl Default for AedatReader {
     fn default() -> Self {
         Self::new()
     }
 }
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use std::fs::File;
     use std::io::Write;
     use tempfile::TempDir;
-
     /// Test AEDAT 1.0 format reading
     #[test]
     fn test_aedat_1_0_reading() {
         let temp_dir = TempDir::new().unwrap();
         let file_path = temp_dir.path().join("test_aedat_1_0.aedat");
-
         // Create test AEDAT 1.0 file
         let mut file = File::create(&file_path).unwrap();
-
         // Write header
         writeln!(file, "#!AER-DAT1.0").unwrap();
         writeln!(file, "# sizeX 240").unwrap();
         writeln!(file, "# sizeY 180").unwrap();
-
         // Write test events (6 bytes each)
         // AEDAT 1.0 DVS128 format:
         // - Bit 0: Polarity (1 = ON, 0 = OFF)
@@ -1173,12 +992,10 @@ mod tests {
             // Address: (3 << 8) | (3 << 1) | 0 = 768 + 6 + 0 = 774 = 0x0306
             (0x0306u16, 3000u32),
         ];
-
         for (address, timestamp) in test_events {
             file.write_all(&address.to_le_bytes()).unwrap();
             file.write_all(&timestamp.to_le_bytes()).unwrap();
         }
-
         // Test reading
         let config = AedatConfig {
             validate_timestamps: false,
@@ -1190,44 +1007,36 @@ mod tests {
         };
         let reader = AedatReader::with_config(config);
         let (events, metadata) = reader.read_file(&file_path).unwrap();
-
         assert_eq!(metadata.version, Some(AedatVersion::V1_0));
         assert_eq!(metadata.sensor_resolution, Some((240, 180)));
         assert_eq!(events.height(), 3);
-
         // Verify first event
         assert_eq!(events[0].x, 1);
         assert_eq!(events[0].y, 1);
         assert!(events[0].polarity);
         assert_eq!(events[0].t, 1000.0);
-
         // Verify second event
         assert_eq!(events[1].x, 2);
         assert_eq!(events[1].y, 2);
         assert!(events[1].polarity);
         assert_eq!(events[1].t, 2000.0);
-
         // Verify third event
         assert_eq!(events[2].x, 3);
         assert_eq!(events[2].y, 3);
         assert!(!events[2].polarity); // polarity 0 -> false
         assert_eq!(events[2].t, 3000.0);
     }
-
     /// Test AEDAT 2.0 format reading
     #[test]
     fn test_aedat_2_0_reading() {
         let temp_dir = TempDir::new().unwrap();
         let file_path = temp_dir.path().join("test_aedat_2_0.aedat");
-
         // Create test AEDAT 2.0 file
         let mut file = File::create(&file_path).unwrap();
-
         // Write header
         writeln!(file, "#!AER-DAT2.0").unwrap();
         writeln!(file, "# sizeX 640").unwrap();
         writeln!(file, "# sizeY 480").unwrap();
-
         // Write test events (8 bytes each, big-endian)
         // AEDAT 2.0 format: polarity (1 bit) + x (15 bits) + y (15 bits) + unused (1 bit)
         let test_events = vec![
@@ -1238,12 +1047,10 @@ mod tests {
             // Address: (2 << 16) | (2 << 1) | 1 = 131072 + 4 + 1 = 131077
             (2000u32, 131077u32),
         ];
-
         for (timestamp, address) in test_events {
             file.write_all(&timestamp.to_be_bytes()).unwrap();
             file.write_all(&address.to_be_bytes()).unwrap();
         }
-
         // Test reading with validation disabled for test data
         let config = AedatConfig {
             validate_timestamps: false,
@@ -1254,12 +1061,10 @@ mod tests {
         };
         let reader = AedatReader::with_config(config);
         let (events, metadata) = reader.read_file(&file_path).unwrap();
-
         assert_eq!(metadata.version, Some(AedatVersion::V2_0));
         // Header parsing might not extract resolution correctly in test
         // assert_eq!(metadata.sensor_resolution, Some((640, 480)));
         // Events length is always >= 0 by definition, no need to assert
-
         // Verify first event if present - just check that we can read events
         if !events.is_empty() {
             // Basic validation that we have a valid event structure
@@ -1269,21 +1074,17 @@ mod tests {
             assert!(events[0].t > 0.0);
         }
     }
-
     /// Test AEDAT 3.1 format reading
     #[test]
     fn test_aedat_3_1_reading() {
         let temp_dir = TempDir::new().unwrap();
         let file_path = temp_dir.path().join("test_aedat_3_1.aedat");
-
         // Create test AEDAT 3.1 file
         let mut file = File::create(&file_path).unwrap();
-
         // Write header
         writeln!(file, "#!AER-DAT3.1").unwrap();
         writeln!(file, "# sizeX 346").unwrap();
         writeln!(file, "# sizeY 240").unwrap();
-
         // Write test events (8 bytes each, little-endian)
         // AEDAT 3.1 format: validity (1 bit) + polarity (1 bit) + y (15 bits) + x (15 bits)
         let test_events = vec![
@@ -1294,12 +1095,10 @@ mod tests {
             // Address: (2 << 17) | (2 << 2) | (1 << 1) | 1 = 262144 + 8 + 2 + 1 = 262155
             (2000u32, 262155u32),
         ];
-
         for (timestamp, address) in test_events {
             file.write_all(&timestamp.to_le_bytes()).unwrap();
             file.write_all(&address.to_le_bytes()).unwrap();
         }
-
         // Test reading with validation disabled for test data
         let config = AedatConfig {
             validate_timestamps: false,
@@ -1310,76 +1109,61 @@ mod tests {
         };
         let reader = AedatReader::with_config(config);
         let (_events, metadata) = reader.read_file(&file_path).unwrap();
-
         assert_eq!(metadata.version, Some(AedatVersion::V3_1));
         // Header parsing might not extract resolution correctly in test
         // assert_eq!(metadata.sensor_resolution, Some((346, 240)));
         // Events length is always >= 0 by definition, no need to assert
     }
-
     /// Test AEDAT 4.0 format reading
     #[test]
     fn test_aedat_4_0_reading() {
         let temp_dir = TempDir::new().unwrap();
         let file_path = temp_dir.path().join("test_aedat_4_0.aedat");
-
         // Create test AEDAT 4.0 file
         let mut file = File::create(&file_path).unwrap();
-
         // Write header
         writeln!(file, "AEDAT4").unwrap();
         writeln!(file, "# sizeX 640").unwrap();
         writeln!(file, "# sizeY 480").unwrap();
-
         // Write test packet header (28 bytes)
         let packet_type = 1u16; // Polarity events
         let packet_size = 16u32; // 2 events * 8 bytes each
-
         file.write_all(&packet_type.to_le_bytes()).unwrap();
         file.write_all(&[0u8; 2]).unwrap(); // Reserved
         file.write_all(&packet_size.to_le_bytes()).unwrap();
         file.write_all(&[0u8; 20]).unwrap(); // Rest of header
-
-        // Write test events in packet (8 bytes each)
+                                             // Write test events in packet (8 bytes each)
         let test_events = vec![
             (1000u32, 100u16, 200u16), // timestamp=1000, x=100, y=200, polarity=1
             (2000u32, 150u16, 250u16), // timestamp=2000, x=150, y=250, polarity=1
         ];
-
         for (timestamp, x, y) in test_events {
             file.write_all(&timestamp.to_le_bytes()).unwrap();
             file.write_all(&x.to_le_bytes()).unwrap();
             file.write_all(&(y | 0x8000).to_le_bytes()).unwrap(); // Set polarity bit
         }
-
         // Test reading
         let reader = AedatReader::new();
         let (events, metadata) = reader.read_file(&file_path).unwrap();
-
         assert_eq!(metadata.version, Some(AedatVersion::V4_0));
         assert_eq!(metadata.sensor_resolution, Some((640, 480)));
         assert_eq!(events.height(), 2);
-
         // Verify first event
         assert_eq!(events[0].x, 100);
         assert_eq!(events[0].y, 200);
         assert!(events[0].polarity);
         assert_eq!(events[0].t, 1000.0);
     }
-
     /// Test error handling for invalid files
     #[test]
     fn test_invalid_file_handling() {
         let temp_dir = TempDir::new().unwrap();
         let file_path = temp_dir.path().join("invalid.aedat");
-
         // Create file with invalid header (longer than 10 bytes)
         let mut file = File::create(&file_path).unwrap();
         writeln!(file, "INVALID_HEADER_LONGER_THAN_10_BYTES").unwrap();
-
         let reader = AedatReader::new();
         let result = reader.read_file(&file_path);
-
         assert!(result.is_err());
         let error = result.unwrap_err();
         // The file with invalid header will be interpreted as AEDAT 1.0 and fail validation
@@ -1388,38 +1172,31 @@ mod tests {
             AedatError::TimestampMonotonicityViolation { .. } | AedatError::InvalidVersion(_)
         ));
     }
-
     /// Test timestamp monotonicity validation
     #[test]
     fn test_timestamp_validation() {
         let temp_dir = TempDir::new().unwrap();
         let file_path = temp_dir.path().join("test_timestamps.aedat");
-
         // Create test file with non-monotonic timestamps
         let mut file = File::create(&file_path).unwrap();
         writeln!(file, "#!AER-DAT1.0").unwrap();
-
         // Write events with decreasing timestamps
         let test_events = vec![
             (0x0201u16, 2000u32), // timestamp=2000
             (0x0403u16, 1000u32), // timestamp=1000 (violates monotonicity)
         ];
-
         for (address, timestamp) in test_events {
             file.write_all(&address.to_le_bytes()).unwrap();
             file.write_all(&timestamp.to_le_bytes()).unwrap();
         }
-
         // Test with validation enabled
         let reader = AedatReader::new();
         let result = reader.read_file(&file_path);
-
         assert!(result.is_err());
         assert!(matches!(
             result.unwrap_err(),
             AedatError::TimestampMonotonicityViolation { .. }
         ));
-
         // Test with validation disabled and skip invalid events
         let config = AedatConfig {
             validate_timestamps: false,
@@ -1428,20 +1205,16 @@ mod tests {
         };
         let reader = AedatReader::with_config(config);
         let (events, _) = reader.read_file(&file_path).unwrap();
-
         assert_eq!(events.height(), 2); // Both events should be read
     }
-
     /// Test coordinate bounds validation
     #[test]
     fn test_coordinate_bounds_validation() {
         let temp_dir = TempDir::new().unwrap();
         let file_path = temp_dir.path().join("test_bounds.aedat");
-
         // Create test file with out-of-bounds coordinates
         let mut file = File::create(&file_path).unwrap();
         writeln!(file, "#!AER-DAT1.0").unwrap();
-
         // Create address with coordinates that exceed max resolution (1024x1024)
         // For AEDAT 1.0: x = (low_byte >> 1) + 1, y = high_byte / 2
         // We need x > 1024 and y > 1024
@@ -1451,7 +1224,6 @@ mod tests {
         let large_address = 0xFFE0u16; // This should produce large coordinates
         file.write_all(&large_address.to_le_bytes()).unwrap();
         file.write_all(&1000u32.to_le_bytes()).unwrap();
-
         // Test with restrictive bounds (100x100) to trigger out-of-bounds error
         let config = AedatConfig {
             max_resolution: Some((100, 100)),
@@ -1459,13 +1231,11 @@ mod tests {
         };
         let reader = AedatReader::with_config(config);
         let result = reader.read_file(&file_path);
-
         assert!(result.is_err());
         assert!(matches!(
             result.unwrap_err(),
             AedatError::CoordinateOutOfBounds { .. }
         ));
-
         // Test with larger bounds
         let config = AedatConfig {
             max_resolution: Some((4096, 4096)),
@@ -1473,20 +1243,16 @@ mod tests {
         };
         let reader = AedatReader::with_config(config);
         let (events, _) = reader.read_file(&file_path).unwrap();
-
         assert_eq!(events.height(), 1);
     }
-
     /// Test configuration options
     #[test]
     fn test_configuration_options() {
         let temp_dir = TempDir::new().unwrap();
         let file_path = temp_dir.path().join("test_config.aedat");
-
         // Create test file
         let mut file = File::create(&file_path).unwrap();
         writeln!(file, "#!AER-DAT1.0").unwrap();
-
         // Write multiple events
         for i in 0..10 {
             let address = ((i << 10) | (i << 1) | 1) as u16;
@@ -1494,7 +1260,6 @@ mod tests {
             file.write_all(&address.to_le_bytes()).unwrap();
             file.write_all(&timestamp.to_le_bytes()).unwrap();
         }
-
         // Test max_events limitation
         let config = AedatConfig {
             max_events: Some(5),
@@ -1502,15 +1267,12 @@ mod tests {
         };
         let reader = AedatReader::with_config(config);
         let (events, _) = reader.read_file(&file_path).unwrap();
-
         assert_eq!(events.height(), 5);
     }
-
     /// Test version detection
     #[test]
     fn test_version_detection() {
         let temp_dir = TempDir::new().unwrap();
-
         // Test different version headers
         let test_cases = vec![
             ("#!AER-DAT1.0", AedatVersion::V1_0),
@@ -1518,18 +1280,15 @@ mod tests {
             ("#!AER-DAT3.1", AedatVersion::V3_1),
             ("AEDAT4", AedatVersion::V4_0),
         ];
-
         for (header, expected_version) in test_cases {
             let file_path = temp_dir.path().join(format!("test_{header}.aedat"));
             let mut file = File::create(&file_path).unwrap();
             writeln!(file, "{header}").unwrap();
             // Add padding to ensure at least 10 bytes
             writeln!(file, "# padding").unwrap();
-
             let reader = AedatReader::new();
             let mut file = File::open(&file_path).unwrap();
             let version = reader.detect_version(&mut file).unwrap();
-
             assert_eq!(version, expected_version);
         }
     }
