+++ src/tracing_config.rs
@@ -240,7 +240,6 @@ pub mod examples {
 }
 
 /// Python bindings for tracing configuration
-#[cfg(feature = "python")]
 pub mod python {
     use pyo3::prelude::*;
